Fix building type filters and UI issues

+++ satisfactor_ui/app.py
@@ -2,6 +2,9 @@
 
 GTK_APP_ID="com.github.ryanjjung.satisfactor_py"
 
+import logging
+logging.basicConfig(level=logging.DEBUG)
+
 import gi
 gi.require_version('Gtk', '4.0')
 
@@ -44,6 +47,7 @@ class MainWindow(Gtk.ApplicationWindow):
     ):
         super().__init__(*args, **kwargs)
 
+        logging.debug('Initiating main window')
         self.factory = None
         self.factoryFile = None
         self.unsaved_changes = False
@@ -67,6 +71,7 @@ class MainWindow(Gtk.ApplicationWindow):
         unblock_all_signals.
         '''
 
+        logging.debug('Blocking all signals')
         for handler in self.windowSignals:
             GObject.signal_handler_block(handler[0], handler[1])
 
@@ -85,6 +90,7 @@ class MainWindow(Gtk.ApplicationWindow):
 
         global ALL_BUILDINGS
         if ALL_BUILDINGS is None:
+            logging.debug('Initializing buildings list')
             ALL_BUILDINGS = [ ResourceNode(), InfiniteSupplyNode() ]
             ALL_BUILDINGS.extend([ bldg() for bldg in get_all_buildings()])
             ALL_BUILDINGS.extend([ bldg() for bldg in get_all_conveyances()])
@@ -99,14 +105,15 @@ class MainWindow(Gtk.ApplicationWindow):
         try:
             # Load the factory and store it in the class as separate actions,
             # resulting in no change if an exception is thrown at load time.
+            logging.debug(f'Loading factory from file {filename}')
             loadedFactory = Factory.load(filename)
             self.factoryFile = filename
             self.factory = loadedFactory
             self.unsaved_changes = False
-            self.update_window()
             self.set_tier_and_upgrade()
+            self.update_window()
         except IOError as ex:
-            print(f'[DEBUG] An error occurred when loading a factory from file {filename}\n  {ex}')
+            logging.error(f'An error occurred when loading a factory from file {filename}\n  {ex}')
             # TODO: Replace with an actual ErrorDialog
 
     def set_tier_and_upgrade(self):
@@ -114,15 +121,19 @@ class MainWindow(Gtk.ApplicationWindow):
         Sets the active values in the combo boxes for tier and upgrade
         '''
 
+        self.block_all_signals()
+        logging.debug(f'Selecting tier/upgrade values: {self.factory.tier}/{self.factory.upgrade}')
         self.cboTier.set_active(self.factory.tier)
         self.__cboTier_changed(self.cboTier)
         self.cboUpgrade.set_active(self.factory.upgrade - 1)
+        self.unblock_all_signals()
 
     def set_window_title(self):
         '''
         Update the window's title appropriately
         '''
 
+        logging.debug('Setting window title')
         title_prefix = f'{"*" if self.unsaved_changes else ""}'
         title_suffix = f' ({self.factoryFile.split('/')[-1]})' if self.factoryFile else ''
         self.set_title(f'{title_prefix}{MAIN_WINDOW_TITLE_BASE}{title_suffix}')
@@ -133,27 +144,37 @@ class MainWindow(Gtk.ApplicationWindow):
         signals were likely placed by block_all_signals.
         '''
 
+        logging.debug('Unblocking all signals')
         for handler in self.windowSignals:
             GObject.signal_handler_unblock(handler[0], handler[1])
 
-    def __update_buildings_list(self):
+    def update_buildings_list(self):
         '''
         Updates the list of buildings in the left panel, taking into account all filters.
         '''
 
         filterAvailability = True
 
-        availTier = int(self.cboTier.get_active())
-        availUpgrade = int(self.cboUpgrade.get_active())
-
-        buildings = self.get_building_options()
+        all_buildings = self.get_building_options()
         if filterAvailability:
-            buildings = [ building for building in buildings
-                if building.availability.tier >= availTier
-                and building.availability.upgrade >= availUpgrade ]
+            logging.debug(
+                f'Updating filtered list of buildings for tier/upgrade '
+                f'{self.factory.tier}/{self.factory.upgrade}')
+            avail_buildings = []
+            for building in all_buildings:
+                if self.factory.tier > building.availability.tier:
+                    avail_buildings.append(building)
+                elif self.factory.tier == building.availability.tier:
+                    if self.factory.upgrade >= building.availability.upgrade:
+                        avail_buildings.append(building)
+        
+        # Sort the list alphabetically
+        avail_buildings = sorted(avail_buildings, key=lambda x: x.name)
 
         listStore = Gtk.ListStore(Pixbuf, str)
-        for building in buildings:
+        ct = 0
+        for building in avail_buildings:
+            ct += 1
             listStore.append((None, building.name))
         self.icovwBuildings.set_model(listStore)
         self.icovwBuildings.set_pixbuf_column(0)
@@ -165,6 +186,7 @@ class MainWindow(Gtk.ApplicationWindow):
         UI elements depending on that context.
         '''
 
+        logging.debug('Updating window')
         if not self.updating:
             self.updating = True
             self.block_all_signals()
@@ -174,7 +196,7 @@ class MainWindow(Gtk.ApplicationWindow):
                 self.boxFactoryFunctions.set_sensitive(True)
                 if not skipFactoryName:
                     self.entryFactoryName.get_buffer().set_text(self.factory.name, -1)
-                self.__update_buildings_list()
+                self.update_buildings_list()
             else:
                 self.btnSaveFactory.set_sensitive(False)
                 self.boxFactoryFunctions.set_sensitive(False)
@@ -192,6 +214,7 @@ class MainWindow(Gtk.ApplicationWindow):
         designer area.
         '''
 
+        logging.debug('Building the main window')
         # Track all signals so we can block/unblock them easily
         self.windowSignals = []
 
@@ -205,7 +228,7 @@ class MainWindow(Gtk.ApplicationWindow):
         # Build the left-hand panel containing the list of buildings
         self.__buildings_options()
         self.paneLeft.set_start_child(self.paneBuildingsOptions)
-        self.paneLeft.set_position(200)
+        self.paneLeft.set_position(300)
         self.paneLeft.set_vexpand(True)
 
         # Build the right-hand panel as another split panel
@@ -236,19 +259,31 @@ class MainWindow(Gtk.ApplicationWindow):
         available to the user.
         '''
 
+        logging.debug('Building widgets for building options')
         # Start with two vertical panes
         self.paneBuildingsOptions = Gtk.Paned(orientation=Gtk.Orientation.VERTICAL)
         self.paneBuildingsOptions.set_position(300)
         self.lblBuildingsFilters = Gtk.Label(label='Filters')
 
         # Bottom pane: list of buildings
+        self.scrollBuildings = Gtk.ScrolledWindow()
         self.icovwBuildings = Gtk.IconView()
+        self.scrollBuildings.set_child(self.icovwBuildings)
+        self.scrollBuildings.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
 
         # Compile panel contents
         self.paneBuildingsOptions.set_start_child(self.lblBuildingsFilters)
-        self.paneBuildingsOptions.set_end_child(self.icovwBuildings)
+        self.paneBuildingsOptions.set_end_child(self.scrollBuildings)
 
     def __connect_handlers(self):
+        '''
+        Connects signals for the widgets on this window. This is done as a separate task after the
+        window has been fully constructed. This prevents signals from being emitted before the
+        window is functional.
+        '''
+
+        logging.debug('Connecting widget signals')
+        
         self.windowSignals.append((
             self.btnNewFactory,
             self.btnNewFactory.connect('clicked', self.__btnNewFactory_clicked)))
@@ -280,6 +315,8 @@ class MainWindow(Gtk.ApplicationWindow):
         '''
         Builds the UI controls which run across the top bar of the window.
         '''
+
+        logging.debug('Building the top bar')
 
         # The bar's top level object is a horizontal box with some padding
         self.boxTopBar = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL)
@@ -350,6 +387,7 @@ class MainWindow(Gtk.ApplicationWindow):
         Builds reusable items which are unique to this application
         '''
 
+        logging.debug('Building resuable UI helpers')
         self.satFileFilter = Gtk.FileFilter()
         self.satFileFilter.set_name('SatisFactories (*.sat)')
         self.satFileFilter.add_pattern('*.sat')
@@ -484,13 +522,19 @@ class MainWindow(Gtk.ApplicationWindow):
         accordingly.
         '''
 
-        self.cboUpgrade.remove_all()
-        for upgrade in Availability.get_upgrade_strings(self.cboTier.get_active()):
-            self.cboUpgrade.append(upgrade, upgrade)
+        if self.factory:
+            self.factory.tier = self.cboTier.get_active()
+            self.cboUpgrade.remove_all()
+            for upgrade in Availability.get_upgrade_strings(self.factory.tier):
+                self.cboUpgrade.append(upgrade, upgrade)
+            self.cboUpgrade.set_active(0)
+            self.unsaved_changes = True
 
 
     # + "Upgrade" combo box signal handlers
     def __cboUpgrade_changed(self, cbo):
+        self.factory.upgrade = self.cboUpgrade.get_active() + 1
+        self.unsaved_changes = True
         self.update_window()
 
 
@@ -500,6 +544,7 @@ class FactoryDesigner(Gtk.Application):
     '''
 
     def __init__(self, **kwargs):
+        logging.debug('Initializing GTK application')
         super().__init__(application_id=APP_ID, **kwargs)
         self.mainWindow = None
         self.set_flags(Gio.ApplicationFlags.HANDLES_OPEN)
